+++ src/Integration/ErrorListenerIntegration.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Sentry\Integration;
 
 use Sentry\ErrorHandler;
+use Sentry\Exception\FatalErrorException;
 use Sentry\Exception\SilencedErrorException;
 use Sentry\Options;
 use Sentry\State\Hub;
@@ -21,13 +22,24 @@ final class ErrorListenerIntegration implements IntegrationInterface
     private $options;
 
     /**
+     * @var bool Whether to handle fatal errors or not
+     */
+    private $handleFatalErrors;
+
+    /**
      * ErrorListenerIntegration constructor.
      *
-     * @param Options $options The options to be used with this integration
+     * @param Options $options           The options to be used with this integration
+     * @param bool    $handleFatalErrors Whether to handle fatal errors or not
      */
-    public function __construct(Options $options)
+    public function __construct(Options $options, bool $handleFatalErrors = true)
     {
         $this->options = $options;
+        $this->handleFatalErrors = $handleFatalErrors;
+
+        if ($handleFatalErrors) {
+            @trigger_error(sprintf('Handling fatal errors with the "%s" class is deprecated since version 2.1. Use the "%s" integration instead.', self::class, FatalErrorListenerIntegration::class), E_USER_DEPRECATED);
+        }
     }
 
     /**
@@ -35,14 +47,20 @@ final class ErrorListenerIntegration implements IntegrationInterface
      */
     public function setupOnce(): void
     {
-        ErrorHandler::addErrorListener(function (\ErrorException $error): void {
-            if ($error instanceof SilencedErrorException && !$this->options->shouldCaptureSilencedErrors()) {
+        ErrorHandler::addErrorListener(function (\ErrorException $exception): void {
+            if (!$this->handleFatalErrors && $exception instanceof FatalErrorException) {
                 return;
             }
 
-            if ($this->options->getErrorTypes() & $error->getSeverity()) {
-                Hub::getCurrent()->captureException($error);
+            if ($exception instanceof SilencedErrorException && !$this->options->shouldCaptureSilencedErrors()) {
+                return;
             }
+
+            if (!($this->options->getErrorTypes() & $exception->getSeverity())) {
+                return;
+            }
+
+            Hub::getCurrent()->captureException($exception);
         });
     }
 }